+++ navigation.js
@@ -1,0 +1,7 @@
+import { createSharedPathnamesNavigation } from 'next-intl/navigation';
+
+export const locales = ['en', 'fr'];
+export const localePrefix = 'as-needed';
+
+export const { Link, redirect, usePathname, useRouter } =
+  createSharedPathnamesNavigation({ locales, localePrefix });
