+++ src/app/[locale]/page.tsx
@@ -1,10 +1,11 @@
 "use client";
 
 import type { NextPage } from "next";
-import styles from "./page.module.css";
 
 import Description from "@/components/description/description";
 import Introduction from "@/components/introduction/introduction";
+
+import styles from "./page.module.css";
 
 const Home : NextPage = () => {
   return (
